Wizard: Improve controls
I've changed the wizard controls API to make it easier to reuse.
This removed a lot of complexity attached to th rendering of buttons texts
are now moved to the consumer.

I've kept the messages for future reference as I'm thinking we might
create a defaultControls component that comes with batteries included.