Refactor state to useReducer()
This finally solves the infitite loop problem because our actions are no
longer need to be "subscribing" to state changes (e.g. [onComplete] is
the only item in our actions dependency list)

This was made possible by a rewrite to a Flux/Redux like pattern using
useThunkReducer (I needed thunk instead of the regular useReducer because
validators can be async). The extra dependency is 1Kb gzipped.

To be honest, I think we could have done this without using useReducer,
however – this pattern forces you to create clear boundaries between the
ways you interact with your state. As a added bonus, it decreases entanglement
(increasing readability) and is easier to test and debug.